requirements.txt
asgiref==3.9.1
attrs==25.3.0
backports-datetime-fromisoformat==2.0.3
black==25.9.0
click==8.3.0
Django==5.2.6
django-debug-toolbar==6.0.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
environs==14.3.0
importlib_resources==6.5.2
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
marshmallow==4.0.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
PyJWT==2.10.1
python-dotenv==1.1.1
pytokens==0.1.10
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
rpds-py==0.27.1
sqlparse==0.5.3
swagger-spec-validator==3.0.4
tomli==2.2.1
typing_extensions==4.15.0
uritemplate==4.2.0
